assignment4/code/GraphCuts/requirements.txt
# python==3.8.*
numpy==1.22.*
matplotlib==3.5.*
scipy==1.8.*
PyMaxflow==1.3.0
